+++ shopify/packages/hybrid-cart/src/utils.ts
@@ -10,8 +10,7 @@ export const uniq = (array: string[]): string[] =>
 export const getNormalizedAttributes = () => {
   const cart = Cart.get();
   const mergedAttributes = {
-    ...(cart?.attributes?.private || {}),
-    ...(cart?.attributes?.public || {}),
+    ...cart.attributes,
   };
   return mergedAttributes;
 };
@@ -25,8 +24,7 @@ export const getCartDiscountCodes = () => {
   const itemLevelCodes = (cart?.items || [])?.reduce((codes, item) => {
     const itemCodes = (item?.line_level_discount_allocations || [])
       .filter(
-        (allocation) =>
-          allocation.discount_application.type === "discount_code",
+        (allocation) => allocation.discount_application.type === "discount_code"
       )
       .map((allocation) => allocation.discount_application.title);
     return uniq([...codes, ...itemCodes]);
@@ -41,7 +39,7 @@ function defaultUnwrap<T>(cart: CartData): T {
 
 export function subscribe<T>(
   unwrapValue: (cart: CartData) => T = defaultUnwrap,
-  cb: (data: T) => void,
+  cb: (data: T) => void
 ): () => void {
   // Assuming cartQuery and observable are defined and set correctly.
   const observer = observable(() => unwrapValue(Cart.query.data));