+++ shopify/packages/hybrid-cart-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-hybrid-cart-components
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-hybrid-cart@0.1.13
+
 ## 0.1.17
 
 ### Patch Changes